Prebuilt: Switch to Magisk for root solution [1/3]
This will add the Magisk Manager into the ROM and move the zip to be
flashed during ROM installation. Solution is stable and actively
supported by its creator. Commit actually written by @nathanchance,
with authorship given to topjohnwu since he did everything that is
being added.

The su binary is removed but the Superuser APK is still needed for
root management.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>